backend/py/requirements.txt
fastapi
uvicorn
pika
paho-mqtt
sqlalchemy
